add some examples to test some cases for the rounding computations

+++ esphome/components/resampler/speaker/resampler_speaker.h
@@ -99,7 +99,7 @@ class ResamplerSpeaker : public Component, public speaker::Speaker {
 
   uint32_t buffer_duration_ms_;
 
-  uint32_t callback_remainder_{0};
+  uint64_t callback_remainder_{0};
 
   int32_t playback_differential_ms_{0};
 };
